* doxygen.cfg: Update to newer version of doxygen. * assembly/toolbox/object-template.cs: Add base object classes to implement Toolbox templates. * assembly/toolbox/window-template.cs: Add window template support. * assembly/Makefile.am: Add new files. * assembly/toolbox/action-button.cs: Add template support. * assembly/toolbox/window.cs (Window.Flags): Change to enum. (class Window.TemplateOffset): Change name to WindowTemplateOffset and move to toolbox/window-template.cs. (Window.Window): Add ctor to construct from window template. * assembly/toolbox/gadget.cs (ComponentType): Change field types from uint to int. * assembly/toolbox/toolbox.cs (TemplateOffset): Change name to ObjectTemplateOffset and move to toolbox/object-template.cs. * test/!TBTemplate{!Run,Messages}: Add app to test template classes. * test/tb-template.cs: Add app test code. * test/Makefile.am: Add new test. * setvars: Add new test.
git-svn-id: svn://svn.riscos.info/gccsdk/trunk@6358 5d7976f4-0b0e-0410-a586-d2bac8e3be52

+++ autobuilder/develop/mono/riscosdotnet/assembly/toolbox/gadget.cs
@@ -38,20 +38,20 @@ namespace riscos
 
 			public static class ComponentType
 			{
-				public const uint ActionButton = 128;
-				public const uint OptionButton = 192;
-				public const uint LabelledBox = 256;
-				public const uint Label = 320;
-				public const uint RadioButton = 384;
-				public const uint DisplayField = 448;
-				public const uint WritableField = 512;
-				public const uint Slider = 576;
-				public const uint Draggable = 640;
-				public const uint PopupMenu = 704;
-				public const uint AdjusterArrow = 768;
-				public const uint NumberRange = 832;
-				public const uint StringSet = 896;
-				public const uint Button = 960;
+				public const int ActionButton = 128;
+				public const int OptionButton = 192;
+				public const int LabelledBox = 256;
+				public const int Label = 320;
+				public const int RadioButton = 384;
+				public const int DisplayField = 448;
+				public const int WritableField = 512;
+				public const int Slider = 576;
+				public const int Draggable = 640;
+				public const int PopupMenu = 704;
+				public const int AdjusterArrow = 768;
+				public const int NumberRange = 832;
+				public const int StringSet = 896;
+				public const int Button = 960;
 			}
 
 			/*! \brief The window/menu containing this gadget.  */
@@ -199,7 +199,7 @@ namespace riscos
 			 * \param [in] ObjectID The Toolbox ID of the object.
 			 * \param [in] cmpID The Toolbox ID of the gadget component.
 			 * \return The type of the given gadget.  */
-			public static uint GetType (uint ObjectID, uint cmpID)
+			public static int GetType (uint ObjectID, uint cmpID)
 			{
 				uint type;
 
@@ -208,7 +208,7 @@ namespace riscos
 												Method.GetType,
 												cmpID,
 												out type));
-				return type;
+				return (int)type;
 			}
 
 			/* Generic functions that implement common methods used by gadgets.
